Make 1000 Miles crash less :^)
Implements a few missing TRAPs that were causing crashes when accessing
dialogs within 1000 Miles. FillOval and InverRoundedRect are part of an
animation that plays when you go to close the About dialog.

GetKeys is called when displaying Game > Show Cards (maybe a secret easter
egg?).

InverRoundedRect uses a non-rounded rect for an initial implementation and
GetKeys does nothing but LOG a warning (not setting anything the VAR theKeys
is equivalent to no keys being pressed at the time of the TRAP).