ResourceManager: Fallback to System file if provided
On classic MacOS the ResourceManager is supposed to look in the System
resource fork if a resource ID can not be found in the applications.

In Cyder it is not feasible to try to reproduce every resource from the
System file (at this time) and since we are already loading _Pack's from
there it makes sense to allow resources to be loaded from there too.