Implement TRAP SysEdit
This TRAP returns false if the currently active window is NOT a desk
accessory -- since we do not support desk accessories we can always
return false!

This is the last TRAP needing to be implemented in order to make 1000
Miles no long crash when clicking any of the menu items.