Add File I/O TRAPs dummy implementations
Added logging only implementations of (sync):
 - Create
 - Open
 - GetFileInfo
 - SetFileInfo
 - SerEOF
 - Allocate
 - Read
 - Write
 - Close

Read/Write are implemented enough to allow a 1000 Miles game to be
saved/opened verifying it works (albeit only for that specific usage and
only as an in-memory store).

To make printing of the HFSParam types work I had to modify codegen to
retain the base type of type aliases and use that when building ostream.