Implement TRAP SetOrigin
SetOrigin sets the local coordinate system as though the upper left most
point were at the origin given i.e. if SetOrigin(-20,-20) is called
then the upper left coordinate is (-20,-20) and something drawn at (0,0)
would be 20 pixels from the left and top of the window frame.

The bounds tracks where (0,0) is located in the window w.r.t. the upper
left corner of the screen so it must be updated to reflect the new
origin as well i.e. in the previous example if the window is 50 pixels
from the left hand side of the screen then the left bound would not be
-50 but -70 (to account for the 20 extra pixels to (0,0) in the local
coordinate system).

Converting from local to global coordinates only takes into account the
bounds (which are updated w.r.t. to the new origin) so no drawing code
has to change. :^)