[graphics] Respect thePort's BitMap
Previously we had a single BitmapImage that represented the entire
screen with memory allocated by it outside of the emulation. This got
the project really far but _technically_ is not accurate as the GrafPort
contains a BitMap with a `base_addr` which can be changed or read by an
application (it is documented and used by MacPaint).

This new approach allocates the screen's memory inside of the emulation
and properly sets the BitMap for each Port. There is still a "screen"
BitmapImage setup in main() which reflects the memory that would be
allocated to the monitor output on a Macintosh but all drawing calls now
load a BitmapImage corresponding to thePort before drawing.

There are definitely optimization which can be made (often we are
loading the GrafPort out of emulated memory multiple times for each
drawing call) but 1000 Miles still plays well enough and the emulation
is more accurate now! :^)