Set `modifiers` correctly for WindowActivate event
The WindowActivate event is supposed to set modifiers bit 0 to 1 if
the window is being activated or 0 if it is being deactivated. We do not
currently handle deactivations correctly in WindowManager but this
change ensures Cyder is sending activate messages (rather than the
deactivate events previously being sent by mistake). This fixes MacPaint
and allows items to be selected in the windows!