Fix broken builds in bin/
Adding ThePortImage() function to BitmapImage caused a once generic
class (used in bin/) to be tied to the emulator by depending on its
memory map and Port functions. Moving the helper to quickdraw.h makes a
lot more sense as it is related to the current port.

This fixes the broken build.