Account for non-32 bit clean applications
This is going to need a lot more work. MacPaint is not 32-bit clean (i.e
only the lower 24-bits are used for the address and the upper byte is
for flags internal to MacOS). Ideally we would be able to switch between
32-bit clean and 24-bit modes and handle it consistently for all
pointers...

For now none of the test applications break with this change and
MacPaint is able to get a little further without crashing so that's a
win.