Fix compilation on macOS
Building under macOS Sequoia resulted in a lot of build errors in Abseil
which were resolved by up-revving the sub-module dependencies to the
latest `master` build.

Also needed to added an include for absl::StrCat in a handful of files
which somehow transitively were getting it but stopped when uprev'd.