Add @include syntax to typegen
This is a first "quick and dirty" pass at adding the ability to include
files in typegen which allows the files to be broken up better and
encourages reuse of types.

I really should have written a proper dependency solver for the includes
in main.py but... this is good enough for now.