Add trap dispatch table codegen
This is not being used but the idea was that we could codegen the
TrapManager's logic for dispatching traps and pushing/popping from the
stack and accessing registers.

I do not have the patience for the full refactor right now so this will
remain on the back burner til I do.