Make SillyBalls work!
This is a Color QuickDraw example and I was curious how well it would
work. While Cyder does not support Color QuickDraw yet lying and then
just drawing in black and white does seem to yield results!

The only major issue is that balls are being drawn outside the window
which seems to imply we are not handling clipping correctly.