Only queue native mouse events when requested
The EventManager is used with both emulated events and non-standard
native events (MOVE) which can be dispatched to the emulated code if
we are not careful. While it does not seem to effect any tested programs
it does pollute the queue with events that need to be popped off and
thrown away.