Allow WaitNextEvent() to pause emulation loop
Added the ability for WaitNextEvent() to actually pause the emulation
loop (with a condition variable) when waiting for the next event instead
of always immediately returning and being polled again.

The version enabled right now respects the sleep but there is a second
version guarded with a #if that stalls indefinitely until an event comes
in -- this is likely to break some applications but I wanted to
experiment with how low run time costs can be.