Execute native functions outside of Musashi emulator
Previously native code was executed _during_ the memory read stage of
instruction execution -- quickly changing the world out from underneath
the instruction while it is being read.

This worked well but made it difficult to support calling back into
emulated code from within a native trap handler which is necessary to
implement some traps. This ended up being a rather large refactor to
make things more testable while adding new support.

Under the new model we end the CPU timeslice when the instruction is
about to be fetched. Musashi executes that instruction to completion but
does not execute any further cycles. The emulator "executes" a NOP and
flags that a native function is meant to be called which happens as
soon as m68k_execute() returns. This allows the emulator to be
reentrant.