Add support to call m68k functions from native
Adds a templated CallFunction<>() method which allows native code to
easily setup the stack with arguments, call an m68k subroutine at a
particular address, and return the result (+ reset the PC back).

This relies on a new mechanism in the Emulator class which allows
multiple functions to be registered to a given address and have them
called in FILO (stack) ordering to facilitate signaling the end of an
emulated function.