Dirty hack to fix MacPaint
All the sources I can find say that GrafPort::clip_region is equal to
the port_rect but... that makes MacPaint stop working. Hardcoding it to
the default screen size makes things work and does not seem to have
negative effects on the couple of programs I tested... will need more
investigation.