Add support for FONT width/offset kerning
Adds support for the Offset/Width Table to be able to calculate the
total width of a glyph on screen and how much to add to the max kerning
to adjust where we begin drawing the glyph inside its width.

Implemented TRAP CharWidth based on this information -- MacPaint tools
look perfect now!