Use Regions in the WindowManger
This "completes" the adoption of Regions in Cyder, properly implementing
the update region instead of using the "painter's algorithm" to draw
overlapping windows. Tested with the full suite and everything looks
great!

Unfortunately this implementation is not 100% faithful to the original -
most notably, the GrafPort regions (clip/visible) are supposed to be in
LOCAL coordinates but I am calculating everything in GLOBAL in this
change. Regardless this feels like a good checkpoint that can be cleaned
up in a subsequent change. :^)